+++ ciftools/binary/encoder.py
@@ -3,9 +3,9 @@ import sys
 from typing import Any, List, Protocol, Union
 
 import numpy as np
-from ciftools.bin.data_types import DataType, DataTypeEnum
-from ciftools.bin.encoded_data import EncodedCIFData
-from ciftools.bin.encoding_types import (
+from ciftools.binary.data_types import DataType, DataTypeEnum
+from ciftools.binary.encoded_data import EncodedCIFData
+from ciftools.binary.encoding_types import (
     ByteArrayEncoding,
     DeltaEncoding,
     EncodingEnun,
