+++ ciftools/binary/encoding_types.py
