+++ ciftools/serialization.py
@@ -1,6 +1,6 @@
 import msgpack
-from ciftools.bin.data import BinaryCIFFile
-from ciftools.bin.writer import BinaryCIFWriter
+from ciftools.binary.data import BinaryCIFFile
+from ciftools.binary.writer import BinaryCIFWriter
 from ciftools.models.data import CIFFile
